+++ fv3config/config/__init__.py
@@ -8,6 +8,13 @@ from .rundir import write_run_directory
 from .alter import enable_restart, set_run_duration
 from .derive import get_n_processes, get_run_duration, get_timestep
 from .nudging import get_nudging_assets, update_config_for_nudging
+from .diag_table import (
+    DiagFileConfig,
+    DiagFieldConfig,
+    DiagTable,
+    Packing,
+    FileFormat,
+)
 
 
 def get_default_config():